moved logging code to pioc_support
